health bars work if i dont go back to the maim menu.
i'm currently using one script to do this and it may need to change to 2 so i can get both player and boss to update right

+++ Assets/Scripts/Player scripts/PlayerAttack.cs
@@ -11,7 +11,7 @@ public class PlayerAttack : MonoBehaviour
         {
             IDamageable boss = collision.gameObject.GetComponent<IDamageable>();
             Weapon w = player.EquipedWeapon.GetComponent<Weapon>();
-            int weaponModifier = w.Damage;
+            float weaponModifier = w.Damage;
             boss.TakeDamage(player.Damage + weaponModifier);
         }
     }
